feat: added api connections and changes made to platform launch menu
