+++ src/app/core/utils/Spinner.tsx
@@ -3,8 +3,6 @@ export function Spinner () {
   return (
     <div className="flex flex-row space-x-4">
       <div className="w-10 h-10 rounded-full animate-spin border border-solid border-[#635FC7] border-t-transparent"></div>
-
     </div>
-
   )
 }
